refactor server:side, organized code base.
