+++ server/routes/groupRoutes.ts
@@ -1,10 +1,8 @@
 import {  Router } from "express";
 import { getAllGroups } from "../controllers/groupControllers/getAllGroups";
+
 const router = Router();
 
 router.get("/",getAllGroups);
-
-
-
 
 export {router as groupRoutes};
