+++ server/routes/statRoutes.ts
@@ -1,5 +1,4 @@
 import {  Router } from "express";
+
 const router = Router();
-
-
 export {router as statRoutes};
